+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - New `--exclude` command line option to exclude source files from mutants generation, matching a glob.
 
+- New: `CARGO_MUTANTS_MINIMUM_TEST_TIMEOUT` sets a minimum timeout for cargo tests, in seconds. This can be used to allow more time on slow CI builders. If unset the default is still 20s.
+
 - Added: A new `mutants.out/debug.log` with internal debugging information.
 
 - Improved: The time for check, build, and test is now shown separately in progress bars and output, to give a better indication of which is taking more time in the tree under test. Also, times are show in seconds with one decimal place, and they are styled more consistently.